@@ -2,111 +2,79 @@
 
 namespace App\Repositories;
 
-use App\Components\CustomStatusCodes;
-use App\Contracts\FilmRepositoryInterface;
+use App\Contracts\Repository\FilmRepositoryInterface;
+use App\Models\Comment;
 use App\Models\Film;
-use App\Http\Services\Application\GetAllFilmsApplicationService;
-use App\Http\Services\Application\GetSlugFilmsApplicationService;
-use App\Http\Services\Application\StoreFilmsApplicationService;
-use App\Http\Services\Application\AssignGenreApplicationService;
-use App\Http\Services\Application\StoreCommentsApplicationService;
-use App\Http\Services\General\GeneralResponseService;
 
 
 class FilmRepository implements FilmRepositoryInterface
 {
-    protected $all_films_application_service;
-    protected $store_film_service;
-    protected $slug_film_application_service;
-    protected $assign_genre_film_application_service;
-    protected $general_response_service;
-    protected $store_comment_service;
+
+    protected Film $film;
 
 
-    public function __construct(
-        GetAllFilmsApplicationService $all_films_application_service,
-        GetSlugFilmsApplicationService $slug_film_application_service,
-        GeneralResponseService $general_response_service,
-        StoreFilmsApplicationService $store_film_service,
-        AssignGenreApplicationService $assign_genre_film_application_service,
-        StoreCommentsApplicationService $store_comment_service
-        
-    ) {
-        $this->all_films_application_service = $all_films_application_service;
-        $this->slug_film_application_service = $slug_film_application_service;
-        $this->general_response_service = $general_response_service;
-        $this->store_film_service = $store_film_service;
-        $this->assign_genre_film_application_service = $assign_genre_film_application_service;
-        $this->store_comment_service = $store_comment_service;
+    /**
+     * __construct function
+     *
+     * @param Film $film
+     * @package FilmRepositoryInterface
+     */
+    public function __construct(Film $film)
+    {
+        $this->film = $film;
     }
 
-
-
-    public function getAllFilms()
+    /**
+     * getAll function
+     *
+     * @return object
+     */
+    public function getAll(): object
     {
-        try {
-
-            $content = $this->all_films_application_service->execute();
-
-
-            return  $this->general_response_service->generateResponse(
-                $content,
-                CustomStatusCodes::getSuccessCode(),
-                CustomStatusCodes::getGenralSuccessMessage(),
-                CustomStatusCodes::getHttpSuccessCode(),
-
-            );
-        } catch (\Exception $ex) {
-
-            return  $this->general_response_service->generateResponse(
-                "",
-                CustomStatusCodes::getFailureCode(),
-                $ex->getMessage() . '-' . $ex->getLine(),
-                CustomStatusCodes::getHttpInernalServerCode()
-
-            );
-        }
+        return $this->film::with(['Genre', 'FilmRatings', 'Comments', 'Countries'])->get();
     }
 
-
-    public function getFilmsBySlugName(string $slug)
+    /**
+     * getBySlug function
+     *
+     * @param string $slug
+     * @return array
+     */
+    public function getBySlug(string $slug)
     {
-
-        try {
-            $content = $this->slug_film_application_service->execute($slug);
-
-
-            return  $this->general_response_service->generateResponse(
-                $content,
-                CustomStatusCodes::getSuccessCode(),
-                CustomStatusCodes::getGenralSuccessMessage(),
-                CustomStatusCodes::getHttpSuccessCode(),
-
-            );
-        } catch (\Exception $ex) {
-            return  $this->general_response_service->generateResponse(
-                "",
-                CustomStatusCodes::getFailureCode(),
-                $ex->getMessage() . '-' . $ex->getLine(),
-                CustomStatusCodes::getHttpInernalServerCode()
-
-            );
-        }
+        return $this->film::whereFilmSlug($slug)->first();
     }
 
-    public function storeFilm(object $validated_requet)
+    /**
+     * getCommentsByFilmId function
+     *
+     * @param integer $id
+     * @return void
+     */
+    public function getCommentsByFilmId(int $id)
     {
-        return $this->store_film_service->execute($validated_requet);
+        return Comment::whereFilmId($id)->get();
     }
 
-
-    public function assignGeneriesToFilm(object $validated_requet)
+    /**
+     * getAllById function
+     *
+     * @param integer $id
+     * @return object
+     */
+    public function getAllById(int $id): object
     {
-        return $this->assign_genre_film_application_service->execute($validated_requet);
+        return $this->film::whereId($id)->first();
     }
 
-    public function addCommentsToFilm(object $validated_requet)
+    /**
+     * storeFilms function
+     *
+     * @param array $data
+     * @return object
+     */
+    public function storeFilms(array $data): object
     {
-        return $this->store_comment_service->execute($validated_requet);
+        return $this->film::Create($data);
     }
 }
