@@ -3,136 +3,66 @@
 namespace App\Http\Controllers;
 
 use App\Components\CustomStatusCodes;
+use App\Contracts\CommentInterface;
 use App\Http\Requests\StoreCommentRequest;
-use App\Http\Services\General\GeneralResponseService;
+use App\Services\General\GeneralResponseService;
 use Illuminate\Http\Request;
-use App\Repositories\FilmRepository;
-use Illuminate\Support\Facades\Validator;
 
 class CommentController extends Controller
 {
 
-    public $repository;
-
-
-
-
-
-    public function __construct(FilmRepository $film)
+    public $service;
+    /**
+     * __construct function
+     *
+     * @param CommentInterface $service
+     * @package CommentService
+     */
+    public function __construct(CommentInterface $service)
     {
-        $this->repository = $film;
+        $this->service = $service;
     }
     /**
-     * Display a listing of the resource.
+     * store function
      *
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return object
      */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Display a all listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function getAllFilms()
-    {
-        return $this->repository->getAllFilms();
-    }
-
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
+    public function store(Request $request): object
     {
         try {
             $validate_request = StoreCommentRequest::ApiValidation($request);
             if ($validate_request->fails()) {
-                    $response['code'] = CustomStatusCodes::GENERAL_VALIDATION_CODE;
-                    $response['message'] = $validate_request->errors()->first();
-                    $response['body'] = [];
-                    $response['http_code'] = CustomStatusCodes::HTTP_BAD_REQUEST;
-                    $response['status'] = CustomStatusCodes::RESPONSE_SUCCESS_FALSE;
+                $response['code'] = CustomStatusCodes::GENERAL_VALIDATION_CODE;
+                $response['message'] = $validate_request->errors()->first();
+                $response['body'] = [];
+                $response['http_code'] = CustomStatusCodes::HTTP_BAD_REQUEST;
+                $response['status'] = CustomStatusCodes::RESPONSE_SUCCESS_FALSE;
             } else {
-                $response = $this->repository->addCommentsToFilm($request);
+                $response = $this->service->addCommentsToFilm($request);
             }
-                return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
         } catch (\Exception $ex) {
 
-            return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE,$ex->getFile().'-' .$ex->getMessage().'-'.$ex->getLine(),  CustomStatusCodes::HTTP_BAD_REQUEST,CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
+            return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, $ex->getFile() . '-' . $ex->getMessage() . '-' . $ex->getLine(),  CustomStatusCodes::HTTP_BAD_REQUEST, CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
         }
     }
 
     /**
-     * Display the specified resource.
+     * getCommentsByFilmId function
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return object
      */
-    public function show($id)
+    public function getCommentsByFilmId(Request $request): object
     {
-        //
-    }
+        try {
+            $response = $this->service->getCommentsByFilmId($request);
 
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+        } catch (\Exception $ex) {
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $slug
-     * @return \Illuminate\Http\Response
-     */
-    public function showBySlug($slug)
-    {
-        return $this->repository->getFilmsBySlugName($slug);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
+            return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, $ex->getFile() . '-' . $ex->getMessage() . '-' . $ex->getLine(),  CustomStatusCodes::HTTP_BAD_REQUEST, CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
+        }
     }
 }
