@@ -2,135 +2,106 @@
 
 namespace App\Http\Controllers;
 
-use App\Components\CustomStatusCodes;
+use App\Contracts\FilmInterface;
+use App\Http\Requests\GetFilmRequestBySlug;
 use App\Http\Requests\StoreFilmRequest;
-use App\Http\Services\General\GeneralResponseService;
+use App\Services\General\GeneralResponseService;
 use Illuminate\Http\Request;
-use App\Repositories\FilmRepository;
-use Illuminate\Support\Facades\Validator;
 
 class FilmController extends Controller
 {
 
-    public $repository;
-
-
-
-
-
-    public function __construct(FilmRepository $film)
-    {
-        $this->repository = $film;
-    }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
+    public $service;
 
     /**
-     * Display a all listing of the resource.
+     * FilmController __construct
      *
-     * @return \Illuminate\Http\Response
+     * @param FilmInterface $film
+     * @package FilmService
      */
-    public function getAllFilms()
+    public function __construct(FilmInterface $film)
     {
-        return $this->repository->getAllFilms();
+        $this->service = $film;
     }
 
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
 
     /**
-     * Store a newly created resource in storage.
+     * getAllFilms function
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return object
      */
-    public function store(Request $request)
+    public function getAllFilms(): object
+    {
+        try {
+            $response = $this->service->getAllFilms();
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+        } catch (\Exception $ex) {
+              GeneralResponseService::createExceptionResponse($ex);
+        }
+    }
+
+
+    /**
+     * store function
+     *
+     * @param Request $request
+     * @return object
+     */
+    public function store(Request $request): object
     {
         try {
             $validate_request = StoreFilmRequest::ApiValidation($request);
             if ($validate_request->fails()) {
-                
-                return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, $validate_request->errors()->first(),  CustomStatusCodes::HTTP_BAD_REQUEST,CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
+                $response = GeneralResponseService::ValidationResponse($validate_request->errors()->first());
             } else {
-                $response = $this->repository->storeFilm($request);
-                $response = GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+                $response = $this->service->storeFilm($request);
             }
-            return $response;
+            return $response = GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
         } catch (\Exception $ex) {
-
-            return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, $ex->getMessage(),  CustomStatusCodes::HTTP_BAD_REQUEST,CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
+              GeneralResponseService::createExceptionResponse($ex);
         }
     }
 
+
+
     /**
-     * Display the specified resource.
+     * showBySlug function
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param [type] $slug
+     * @return object
      */
-    public function show($id)
+    public function showBySlug($slug): object
     {
-        //
+        try {
+            $request['film_slug']=$slug;
+            $validate_request = GetFilmRequestBySlug::ApiValidation($request);
+            if ($validate_request->fails()) {
+                $response = GeneralResponseService::ValidationResponse($validate_request->errors()->first());
+            } else {
+                $response = $this->service->getFilmsBySlugName($slug);
+            }
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+        } catch (\Exception $ex) {
+              GeneralResponseService::createExceptionResponse($ex);
+        }
     }
 
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $slug
-     * @return \Illuminate\Http\Response
-     */
-    public function showBySlug($slug)
-    {
-        return $this->repository->getFilmsBySlugName($slug);
-    }
+
 
     /**
-     * Show the form for editing the specified resource.
+     * getFilmCountries function
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return object
      */
-    public function edit($id)
+    public function getFilmCountries(): object
     {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
+        try {
+            $response = $this->service->getFilmCountries();
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+        } catch (\Exception $ex) {
+              GeneralResponseService::createExceptionResponse($ex);
+        }
     }
 }
