@@ -2,126 +2,64 @@
 
 namespace App\Http\Controllers;
 
-use App\Components\CustomStatusCodes;
-use App\Http\Requests\StoreFilmRequest;
+use App\Contracts\GenreServiceInterface;
+use App\Exceptions\GeneralException;
 use App\Http\Requests\StoreGenreRequest;
-use App\Http\Services\General\GeneralResponseService;
+use App\Services\General\GeneralResponseService;
+use Exception;
 use Illuminate\Http\Request;
-use App\Repositories\FilmRepository;
-use Illuminate\Support\Facades\Validator;
 
 class GenreController extends Controller
 {
 
-    public $repository;
+    public $service;
 
-
-
-
-
-    public function __construct(FilmRepository $film)
+    /**
+     * __construct function
+     *
+     * @param GenreServiceInterface $service
+     */
+    public function __construct(GenreServiceInterface $service)
     {
-        $this->repository = $film;
+        $this->service = $service;
     }
+
+
+
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     * Note:- Due to shortage of time i am not going to create new repository for genre
-     * 
-     * 
-     * 
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
+    public function list(): object
     {
         try {
-            $validate_request = StoreGenreRequest::ApiValidation($request);
-            if ($validate_request->fails()) {
-                return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, $validate_request->errors()->first(),  CustomStatusCodes::HTTP_BAD_REQUEST,CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
-            } else {
-                $response = $this->repository->assignGeneriesToFilm($request);
-                $response = GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
-            }
-            return $response;
+            $response = $this->service->listGenres();
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
         } catch (\Exception $ex) {
-            return  GeneralResponseService::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, $ex->getMessage(),  CustomStatusCodes::HTTP_BAD_REQUEST,CustomStatusCodes::RESPONSE_SUCCESS_FALSE);
+            return GeneralResponseService::createExceptionResponse($ex);
         }
     }
 
     /**
-     * Display the specified resource.
+     * Undocumented function
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return void
      */
-    public function show($id)
+    public function store(Request $request): object
     {
-        //
-    }
-
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $slug
-     * @return \Illuminate\Http\Response
-     */
-    public function showBySlug($slug)
-    {
-        return $this->repository->getFilmsBySlugName($slug);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
+        try {
+            $validate_request = StoreGenreRequest::ApiValidation($request);
+            if ($validate_request->fails()) {
+                $response = GeneralResponseService::ValidationResponse($validate_request->errors()->first());
+            } else {
+                $response = $this->service->assignGeneriesToFilm($request);
+            }
+            return GeneralResponseService::responseGenerator($response['body'], $response['code'], $response['message'], $response['http_code'], $response['status']);
+        } catch (\Exception $ex) {
+            return  GeneralResponseService::createExceptionResponse($ex);
+        }
     }
 }
