@@ -2,29 +2,24 @@
 
 namespace App\Http\Requests;
 
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Foundation\Http\FormRequest;
 
 class GetFilmRequestBySlug extends FormRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return false;
-    }
+
 
     /**
      * Get the validation rules that apply to the request.
      *
      * @return array
      */
-    public function rules()
+    public static function ApiValidation($request)
     {
-        return [
-            //
-        ];
+        $validator = Validator::make($request, [
+            'film_slug'     => 'required|exists:films,film_slug|max:100',
+        ]);
+
+        return $validator;
     }
 }
