@@ -12,11 +12,9 @@ class LoginUserRequest
 
     public static function ApiValidation($request)
     {
-        $validator = Validator::make($request->all(), [
+        return Validator::make($request->all(), [
             'email'     => 'required|email|max:50',
             'password'     => 'required|string|regex:/[a-z]/|regex:/[A-Z]/|regex:/[0-9]/|regex:/[@$!%*#?&]/|min:8',
         ]);
-
-        return $validator;
     }
 }
