@@ -16,11 +16,11 @@ class StoreFilmRequest
             'name'     => 'required|max:50',
             'film_slug'     => 'required|unique:films,film_slug|max:100',
             'description'     => 'required',
-            'release_date'     => 'required||max:50',
+            'release_date'     => 'date_format:Y-m-d|required||max:50',
             'ticket_price'     => 'required||max:50',
-            'country'     => 'required||max:100',
-            'photo'     => 'required|file',
-            'genre_id'     => 'required|exists:genres,id|max:50',
+            'country_id'     => 'required||max:100',
+            'photo'     => 'required|file|max:10000',
+            'genre_ids'     => 'required|exists:genres,id|max:50',
         ]);
 
         return $validator;
