@@ -1,25 +1,34 @@
 <?php
 
-namespace App\Http\Services\Domain;
+namespace App\Services\Domain;
 
-use App\Contracts\AbstractGenres;
-use App\Http\Services\Domain\Contracts\DomainServiceInterface;
+use App\Contracts\Repository\FilmGenreRepositoryInterface;
+use App\Contracts\Repository\FilmRepositoryInterface;
+use App\Services\Domain\Contracts\DomainServiceInterface;
 
-class AssignFilmGenreDomainService extends AbstractGenres implements DomainServiceInterface
+class AssignFilmGenreDomainService implements DomainServiceInterface
 {
+    protected FilmGenreRepositoryInterface $repository;
 
-
-
-    public function execute($request)
+    /**
+     * __construct function
+     *
+     * @param FilmGenreRepositoryInterface $repository
+     */
+    public function __construct(FilmGenreRepositoryInterface $repository)
     {
+        $this->repository = $repository;
+    }
 
-
-        $data = $this->assignGenreToFilm($request);
-
-        if ($data == true) {
-            return $data;
-        } else {
-            return false;
-        }
+    /**
+     * execute function
+     *
+     * @param object|array $request
+     * @return object|boolean
+     */
+    public function execute(object|array $request) : object|bool
+    {
+        $data = $this->repository->assignGenreToFilm($request);
+        return $data ? $data :  false;
     }
 }
