@@ -1,19 +1,29 @@
 <?php
 
-namespace App\Http\Services\Domain;
+namespace App\Services\Domain;
 
-use App\Contracts\AbstractFilms;
-use App\Http\Services\Domain\Contracts\DomainServiceInterface;
+use App\Contracts\Repository\FilmRepositoryInterface;
+use App\Services\Domain\Contracts\DomainServiceInterface;
 
-class StoreFilmDomainService extends AbstractFilms implements DomainServiceInterface
+class StoreFilmDomainService  implements DomainServiceInterface
 {
 
 
+    protected FilmRepositoryInterface $repository;
 
-    public function execute($request)
+    public function __construct(FilmRepositoryInterface $repository)
+    {
+        $this->repository = $repository;
+    }
+
+    /**
+     * execute function
+     *
+     * @param object|array $request
+     * @return object|boolean
+     */
+    public function execute(object|array $request) : object|bool
     {   
-
-        //becuaase it is our write layer so all write work will done here
         $photo_move=$request['photo_object'];
         $destination_path=$request['destination_path'];
         $image_name=$request['photo'];
@@ -21,10 +31,11 @@ class StoreFilmDomainService extends AbstractFilms implements DomainServiceInter
         if($photo_move==true){
             unset($request['photo_object']);
             unset($request['destination_path']);
-            return $this->storeFilms($request);
+            return $this->repository->storeFilms($request);
         }else{
             return false;
         }
         
     }
+
 }
