@@ -1,18 +1,16 @@
 <?php
 
-namespace App\Http\Services\General;
+namespace App\Services\General;
 
 use App\Components\CustomStatusCodes;
-use App\Contracts\AbstractFilms;
+use App\Contracts\GeneralResponseServiceInterface;
 
 class GeneralResponseService implements GeneralResponseServiceInterface
 {
 
-
-    public static function generateResponse($body, $code, $message ,$http_code)
+    
+    public static function generateResponse($body, $code, $message, $http_code): object
     {
-
-
         $header['code'] = $code;
         $header['message'] = $message;
 
@@ -20,30 +18,28 @@ class GeneralResponseService implements GeneralResponseServiceInterface
             'header' => $header,
             'body' => $body
         );
-        return response($data,$http_code);
+        return response($data, $http_code);
     }
 
-    public static function successResponseFetch(array $body)
+    public static function successResponseFetch(array $body): object
     {
-        $code= CustomStatusCodes::HTTP_SUCCESS_CODE;
-        $message=CustomStatusCodes::SUCCESS_GENERAL_MESSAGE;
-        $http_code=CustomStatusCodes::HTTP_SUCCESS_CODE;
-        
-        return self::responseGenerator($body,$code,$message,$http_code,true);
+        $code = CustomStatusCodes::HTTP_SUCCESS_CODE;
+        $message = CustomStatusCodes::SUCCESS_GENERAL_MESSAGE;
+        $http_code = CustomStatusCodes::HTTP_SUCCESS_CODE;
 
+        return self::responseGenerator($body, $code, $message, $http_code, true);
     }
 
-    public static function successResponseCreated($body)
+    public static function successResponseCreated($body): object
     {
-        $code= CustomStatusCodes::HTTP_SUCCESS_CODE;
-        $message=CustomStatusCodes::SUCCESS_CREATED;
-        $http_code=CustomStatusCodes::HTTP_INSERTED_SUCCESS_CODE;
-        
-        return self::responseGenerator($body,$code,$message,$http_code,true);
+        $code = CustomStatusCodes::HTTP_SUCCESS_CODE;
+        $message = CustomStatusCodes::SUCCESS_CREATED;
+        $http_code = CustomStatusCodes::HTTP_INSERTED_SUCCESS_CODE;
 
+        return self::responseGenerator($body, $code, $message, $http_code, true);
     }
 
-    public static function responseGenerator($body, $code, $message ,$http_code,$status=true)
+    public static function responseGenerator($body, $code, $message, $http_code, $status = true) : object
     {
         $header['code'] = $code;
         $header['message'] = $message;
@@ -52,9 +48,27 @@ class GeneralResponseService implements GeneralResponseServiceInterface
             'header' => $header,
             'body' => $body
         );
-        return response($data,$http_code);
+        return response($data, $http_code);
     }
 
-    
+    public static function createExceptionResponse($ex): object
+    {
 
+        return self::responseGenerator([], CustomStatusCodes::GENERAL_VALIDATION_CODE, self::GenerateMessageByException($ex), CustomStatusCodes::HTTP_INTERNAL_SERVER_ERROR_CODE, false);
+    }
+
+    public static function GenerateMessageByException(object $ex): string
+    {
+        return $ex->getMessage() . '-' . $ex->getFile() . '-' . $ex->getLine();
+    }
+
+    public static function ValidationResponse($message): array
+    {
+        $response['code'] = CustomStatusCodes::GENERAL_VALIDATION_CODE;
+        $response['message'] = $message;
+        $response['body'] = [];
+        $response['http_code'] = CustomStatusCodes::HTTP_BAD_REQUEST;
+        $response['status'] = CustomStatusCodes::RESPONSE_SUCCESS_FALSE;
+        return $response;
+    }
 }
