@@ -27,8 +27,8 @@ class FilmsTableSeeder extends Seeder
                         'description'=>'random description of table',
                         'release_date'=>'2022-01-01',
                         'ticket_price'=>'30',
-                        'country'=>'UK',
-                        'photo'=>'photo_linlk',
+                        'country_id'=>1,
+                        'photo'=>'https://livijumpa2.files.wordpress.com/2013/12/impossible.jpg',
                         'status'=>'1'
                     
                     ),
@@ -38,8 +38,8 @@ class FilmsTableSeeder extends Seeder
                         'description'=>'random description of table',
                         'release_date'=>'2022-01-01',
                         'ticket_price'=>'40',
-                        'country'=>'US',
-                        'photo'=>'photo_linlk',
+                        'country_id'=>2,
+                        'photo'=>'https://dcassetcdn.com/design_img/16298/35582/35582_406182_16298_image.jpg',
                         'status'=>'1'
                 
                     ),
@@ -49,8 +49,8 @@ class FilmsTableSeeder extends Seeder
                     'description'=>'random description of table',
                     'release_date'=>'2022-01-01',
                     'ticket_price'=>'50',
-                    'country'=>'Germany',
-                    'photo'=>'photo_linlk',
+                    'country_id'=>3,
+                    'photo'=>'https://webneel.com/sites/default/files/imagecache/660Thumbnail/images/blog/movie-poster-design.jpg',
                     'status'=>'1'
             
                 ),
