@@ -21,16 +21,21 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
+Route::get('country/list', [FilmController::class, 'getFilmCountries']);
 
-Route::post('film', [FilmController::class, 'index']);
 Route::get('film/list', [FilmController::class, 'getAllFilms']);
 Route::get('film/{slug}', [FilmController::class, 'showBySlug']);
-Route::post('user/register', [UserController::class, 'store']);
-Route::post('user/login', [UserController::class, 'login']);
 Route::post('film/store', [FilmController::class, 'store']);
 
+
+Route::post('user/register', [UserController::class, 'store']);
+Route::post('user/login', [UserController::class, 'login']);
+Route::get('comment/{film_id}', [CommentController::class, 'getCommentsByFilmId']);
+
 Route::middleware('auth:api')->group(function () {
-    Route::post('genre/store', [GenreController::class, 'store']);
+    Route::post('genre/store', [GenreController::class, 'store']);    
     Route::post('comment/store', [CommentController::class, 'store']);
 });
+Route::get('genre/list', [GenreController::class, 'list']);
+
 
